requirements.txt
telethon
cryptg
motor 
python-dotenv